Move onLogin set state=PLAY to client/play
